+++ mypy/build.py
@@ -344,6 +344,7 @@ class CacheMeta(NamedTuple):
     version_id: str  # mypy version for cache invalidation
     ignore_all: bool  # if errors were ignored
     plugin_data: Any  # config data from plugins
+    baseline_hash: str | None  # hash representing the baselined errors of this module
 
 
 # NOTE: dependencies + suppressed == all reachable imports;
@@ -382,6 +383,7 @@ def cache_meta_from_dict(meta: dict[str, Any], data_json: str) -> CacheMeta:
         meta.get("version_id", sentinel),
         meta.get("ignore_all", True),
         meta.get("plugin_data", None),
+        meta.get("baseline_hash", None),
     )
 
 
@@ -1105,7 +1107,7 @@ def save_baseline(manager: BuildManager):
         manager.stdout.write(f"Baseline successfully updated at {file}\n")
 
 
-def load_baseline(options: Options, errors: Errors, stdout: TextIO) -> None:
+def load_baseline(options: Options, errors: Errors, stdout: TextIO):
     if not options.baseline_file:
         return
 
@@ -1485,6 +1487,25 @@ def find_cache_meta(id: str, path: str, manager: BuildManager) -> CacheMeta | No
         manager.log(f"Metadata abandoned for {id}: plugin configuration differs")
         return None
 
+    baseline_hash = None
+    baseline_errors = manager.errors.baseline.get(path)
+    if baseline_errors:
+        errors = [
+            {
+                "code": error["code"],
+                "column": error["column"],
+                "line": error["line"],
+                "message": error["message"],
+            }
+            for error in baseline_errors
+        ]
+        baseline_hash = hash_digest(json.dumps(errors).encode())
+    if baseline_hash != m.baseline_hash:
+        manager.log(f"Metadata abandoned for {id}: baseline is different")
+        return None
+    if baseline_errors:
+        manager.errors.all_errors[path] = "fresh"
+
     manager.add_stats(fresh_metas=1)
     return m
 
@@ -1651,6 +1672,7 @@ def write_cache(
     source_hash: str,
     ignore_all: bool,
     manager: BuildManager,
+    baseline_errors: list[ErrorInfo] | None,
 ) -> tuple[str, CacheMeta | None]:
     """Write cache files for a module.
 
@@ -1736,6 +1758,19 @@ def write_cache(
         manager.log(f"Error in os.stat({data_json!r}), skipping cache write")
         return interface_hash, None
 
+    baseline_hash = None
+    if baseline_errors:
+        errors = [
+            {
+                "code": error.code and error.code.code,
+                "column": error.column,
+                "line": error.line,
+                "message": error.message,
+            }
+            for error in baseline_errors
+        ]
+        baseline_hash = hash_digest(json.dumps(errors).encode())
+
     mtime = 0 if bazel else int(st.st_mtime)
     size = st.st_size
     # Note that the options we store in the cache are the options as
@@ -1761,6 +1796,7 @@ def write_cache(
         "version_id": manager.version_id,
         "ignore_all": ignore_all,
         "plugin_data": plugin_data,
+        "baseline_hash": baseline_hash,
     }
 
     # Write meta cache file
@@ -2609,7 +2645,7 @@ class State:
 
         return valid_refs
 
-    def write_cache(self) -> None:
+    def write_cache(self, *, baseline_errors: list[ErrorInfo] | None) -> None:
         assert self.tree is not None, "Internal error: method must be called on parsed file only"
         # We don't support writing cache files in fine-grained incremental mode.
         if (
@@ -2648,6 +2684,7 @@ class State:
             self.source_hash,
             self.ignore_all,
             self.manager,
+            baseline_errors,
         )
         if new_interface_hash == self.interface_hash:
             self.manager.log(f"Cached module {self.id} has same interface")
@@ -3599,12 +3636,17 @@ def process_stale_scc(graph: Graph, scc: list[str], manager: BuildManager) -> No
         for id in stale:
             graph[id].transitive_error = True
     for id in stale:
-        if graph[id].xpath not in manager.errors.ignored_files:
-            errors = manager.errors.file_messages(
-                graph[id].xpath, formatter=manager.error_formatter
-            )
-            manager.flush_errors(manager.errors.simplify_path(graph[id].xpath), errors, False)
-        graph[id].write_cache()
+        file = graph[id].xpath
+        if file not in manager.errors.ignored_files:
+            errors = manager.errors.file_messages(file, formatter=manager.error_formatter)
+            manager.flush_errors(manager.errors.simplify_path(file), errors, False)
+        all_errors = manager.errors.all_errors.get(file)
+        if all_errors == "fresh":
+            raise ValueError("huh?")
+        if file in manager.errors.ignored_files:
+            all_errors = None
+        # here, the `all_errors` will only be cached if not `transitive_error`, so `all_errors` is baselined errors
+        graph[id].write_cache(baseline_errors=all_errors)
         graph[id].mark_as_rechecked()